+++ Halesworth/MHSystems/app/src/main/java/com/mh/systems/halesworth/adapter/TabsAdapter/TabsPageAdapter.java
@@ -13,6 +13,7 @@ import android.support.v4.app.FragmentStatePagerAdapter;
 import com.mh.systems.halesworth.activites.BaseActivity;
 import com.mh.systems.halesworth.activites.YourAccountActivity;
 import com.mh.systems.halesworth.constants.ApplicationGlobal;
+import com.mh.systems.halesworth.fragments.ContactUsFragment;
 import com.mh.systems.halesworth.fragments.EligibleFriendsFragment;
 import com.mh.systems.halesworth.fragments.EligibleMemberFragment;
 import com.mh.systems.halesworth.fragments.FinanceFragment;
@@ -133,6 +134,7 @@ public class TabsPageAdapter extends FragmentStatePagerAdapter {
      * Load MEMBERS Tabs i.e
      * <br> 1. Members
      * <br> 2. Friends
+     * <br> 3. Contact Us
      * <p/>
      *
      * @param iPosition
@@ -148,6 +150,10 @@ public class TabsPageAdapter extends FragmentStatePagerAdapter {
             case 1:
                 FriendsFragment friendsFragment = new FriendsFragment();
                 return friendsFragment;
+
+            case 2:
+                ContactUsFragment contactUsFragment = new ContactUsFragment();
+                return contactUsFragment;
 
             default:
                 return null;
@@ -175,7 +181,7 @@ public class TabsPageAdapter extends FragmentStatePagerAdapter {
 
             case 1:
                 fragment = new HandicapFragment();
-               break;
+                break;
 
             case 2:
                 fragment = new FinanceFragment();