+++ Hartsbourne/MHSystems/app/src/main/java/com/mh/systems/hartsbourne/activites/DashboardActivity.java
@@ -15,7 +15,6 @@ import android.widget.LinearLayout;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import com.facebook.drawee.backends.pipeline.Fresco;
 import com.google.gson.JsonObject;
 import com.mh.systems.hartsbourne.R;
 import com.mh.systems.hartsbourne.adapter.RecyclerAdapter.DashboardRecyclerAdapter;
@@ -117,7 +116,6 @@ public class DashboardActivity extends BaseActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        Fresco.initialize(DashboardActivity.this); //Initialize facebook Fresco for round profile pic.
         setContentView(R.layout.activity_dashboard);
 
         /**