+++ MHS Demo/MHSystems/app/src/main/java/com/mh/systems/demoapp/activites/ClubNewsActivity.java
@@ -13,7 +13,6 @@ import android.widget.LinearLayout;
 import android.widget.RelativeLayout;
 import android.widget.TextView;
 
-import com.facebook.drawee.backends.pipeline.Fresco;
 import com.google.gson.JsonObject;
 import com.google.gson.reflect.TypeToken;
 import com.mh.systems.demoapp.R;
@@ -112,7 +111,6 @@ public class ClubNewsActivity extends BaseActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        Fresco.initialize(ClubNewsActivity.this);
         setContentView(R.layout.activity_club_news);
 
         //Initialize view resources.